Add rudimentary mock filesystem object
This is needed for unit testing to correctly simulate different
environments, as Zef is mostly filesystem based. The mock object
currently does not support anything but lfs.attributes with aname ==
'mode' and io.open() functionality, with file:read('*all') being the
only type of file reading supported. (It also can't mock different
attributes on files, timestamps etc, but this has to be changed.) It can
also simulate errors on open and read, for unit testing this behavior.