Check default values for options in Zef.yaml
Since Zef.yaml gives the user the opportunity to set default options, we
should check if these match the type required. So far, this checking has
been offloaded to the stage where ZefConfig.yaml was parsed and a
particular option defaulted to this value, which is later determined to
be invalid. This is unintuitive and adds a big surface for mistakes when
writing the Zef.yaml files and realizing the mistakes are there only
when a particular option is not supplied with a value (so the default is
assumed).

Also add matching tests. This now marks 100% code path coverage in
zef-project, as reported by luacov.