Do not use #table when holes might be present
The value of #table is implementation-defined if the table is not a
valid array i.e. has holes in it. This does not affect vanilla Lua but
manifests itself when running with LuaJIT, as it returns an invalid
'array length', which causes some values in Zef.yaml and ZefConfig to be
incorrectly parsed. Also add some clarification to relevant test
descriptions.