Add a new prototype in MoonScript
This is my vision of how a build script should look, and serves as a
reminder and a unit test later---Zef should be able to build it one day
with no changes, ideally. (This ideal situation will probably not occur,
though.) Also some comments detailing the public-facing interface of
Zef.